requirements.txt
beautifulsoup4==4.13.4
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.0
rsa==4.9.1
soupsieve==2.7
Telethon==1.40.0
typing_extensions==4.13.2
